Dev malina

+++ src/main/java/unibuc/SkillLink/controllers/ClientsController.java
@@ -10,8 +10,10 @@ import unibuc.SkillLink.abstractions.IMediator;
 import unibuc.SkillLink.commands.GetCurrentUserCommand;
 import unibuc.SkillLink.commands.clients.*;
 import unibuc.SkillLink.commands.providers.GetClientsCommand;
+import unibuc.SkillLink.models.AppUser;
 import unibuc.SkillLink.models.Client;
 
+import java.util.Optional;
 import java.util.UUID;
 
 @Controller
@@ -20,10 +22,10 @@ public class ClientsController {
     IMediator mediator;
 
     @PostMapping("/provider/add")
-    public String addProvider(@RequestParam("clientUsername") String clientUsername, @RequestParam("providerUsername") String providerUsername, Model model) {
+    public String addProvider(@RequestParam("clientUsername") String clientUsername, @RequestParam("providerUsername") String providerUsername, Model model, Authentication auth) {
         var client = mediator.handle(new AddProviderCommand(clientUsername, providerUsername));
-        model.addAttribute("client", client);
-        return "redirect:/client/profile";
+        model.addAttribute("user", client);
+        return "redirect:/profile";
     }
 
     @GetMapping("/clients")
@@ -36,8 +38,8 @@ public class ClientsController {
     @GetMapping("/client/{id}")
     public String getClient(@PathVariable UUID id, Model model) {
         var client = mediator.handle(new GetClientCommand(id));
-        model.addAttribute("client", client);
-        return "client/profile";
+        model.addAttribute("user", client);
+        return "client/public-profile";
     }
 
     @GetMapping("client/create")
@@ -54,15 +56,31 @@ public class ClientsController {
     }
 
     @DeleteMapping("/client/{id}")
-    public String deleteClient(@PathVariable UUID id, Model model) {
+    public String deleteClient(@PathVariable UUID id, Model model, Authentication auth) {
+        AppUser currentUser = mediator.handle(new GetCurrentUserCommand(auth));
+
+        if (!(currentUser instanceof Client typeClient) || !typeClient.getId().equals(id)) {
+            model.addAttribute("status", 403);
+            model.addAttribute("message", "Operation not permitted");
+            return "error";
+        }
         mediator.handle(new DeleteClientCommand(id));
-         return "redirect:/index";
+//         return "redirect:/index";
+        return "index";
     }
 
     @PutMapping("/client/{id}")
-    public String editClient(@PathVariable UUID id, @ModelAttribute Client client, Model model) {
+    public String editClient(@PathVariable UUID id, @ModelAttribute Client client, Model model, Authentication auth) {
+        AppUser currentUser = mediator.handle(new GetCurrentUserCommand(auth));
+
+        if (!(currentUser instanceof Client typeClient) || !typeClient.getId().equals(id)) {
+            model.addAttribute("status", 403);
+            model.addAttribute("message", "Operation not permitted");
+            return "error";
+        }
+
         var editedClient = mediator.handle(new EditClientCommand(client));
-        model.addAttribute("client", editedClient);
-        return "client/profile";
+        model.addAttribute("user", editedClient);
+        return "redirect:/profile";
     }
 }
